Move component DataSelector from 'components' to 'components/index/_shared/charts/_base'
